Update araxxor.txt
replaced the experienced preset with my melee setup (if anyone has an undyed sgb feel free to remake it). Cleaned up the guide in general, noticed it is missing quite a bit of emojis but i'm not sure how to add them. will probably come back later to do a writeup of advanced strategies with intent to earn rax master from them