Update araxxor.txt
added advanced strategy guide that should be suitable to get anyone rax master on all paths. worked on it with piegood and helwyr pleb. fixed a typo and replaced p4 zerk rotation video with p4 zgs video. I dont know how to work the table of contents so if someone wants to add the advanced guide and example videos sections to this that would be great