Update araxxor.txt (#68)
* Update araxxor.txt

added advanced strategy guide that should be suitable to get anyone rax master on all paths. worked on it with piegood and helwyr pleb. fixed a typo and replaced p4 zerk rotation video with p4 zgs video. I dont know how to work the table of contents so if someone wants to add the advanced guide and example videos sections to this that would be great

* Update araxxor.txt

path 1 and path 3 advanced guides were >2000 characters so trying to break them into a couple messages

* Update araxxor.txt

* Add table of contents to araxxor guide

* Fix space in quake emoji

* Fixed some formatting issues

* Remove next emoji in favour of unicode arrow

* Embed imgur links

* Update araxxor.txt

Typo fixes for Araxxor

* Magister typo fix

Co-authored-by: Ryyhen <62208155+ryyhen@users.noreply.github.com>
Co-authored-by: hoobsie <59879270+hoobsie@users.noreply.github.com>