Update revo-to-full-manual.txt (#244)
* Update revo-to-full-manual.txt

still early on in this, so very much a work in progress. not ready to upload!! trying to do a full rewrite to be in line with the style guide, and also offer methods of switching from revo to full manual (i.e. can enable queueing with revo to still have revo as a backup but start getting into manually inputting abilities without fighting revo too hard, or ripping off the bandaid and just going head first into full manual) as well as offering safe places to practice full manual like dummies, gwd2, magister, etc. should probably also refer people to #interface-guide as that has the keybind guide in it

* Update revo-to-full-manual.txt

made a bit more progress but stopped at "More advanced scenarios". Will continue working on it when I have time

* Update revo-to-full-manual.txt

* Update revo-to-full-manual.txt

* Update revo-to-full-manual.txt

* Update revo-to-full-manual.txt

* Update revo-to-full-manual.txt