Update perk-info.txt
Lines 3, 104: Improved sentence structure
Lines 12, 32: Substituted "flank" for "flanking", as weapons with the flanking perk on them are generally referred to as a flanking switch rather than a flank switch. 
Lines 20, 39, 53, 70, 83, 99: Fixed grammatical errors
Lines 25,39, 47,65, 104: Fixed punctuation errors
Lines 72, 92: Substituted "us" for "you" , as majority of PVME writeups are written from the perspective of a person referring to the reader, instead of a person referring to themselves and the reader.