Foreword on Power armour vs Tank armour, as it wasn't mentioned anywhere in the channel. Added some notes in Achto explaining that, when it is used, it is often just boots, gloves and helm and why.
The big bit, split out the weapons section to add a T70, T75-T79, T80-T85, T87-T88 sections.
This has nearly doubled the length of the channel.
I think I hit all viable options for training/switches later if needed.

Adjusted most cases of 'completely useless' to 'The special is completely useless' otherwise it reads like we're dismissing the weapon... Then prioritising it as `high`.
Fixed 1 or 2 minor typos.
It may flag up as a lot of changes due to having to merge in later changes at the end, was not the intention!