Update armour-and-weapons.txt
Added Scripture of Wen, forgot in my last commit. Changed Scripture of Jas description to clarify that the final target is whatever you're targeting when it goes off, not the target that originally procced it, and cleaned up the section to clarify that Scriptures are usually meta when not using a Grimoire. Changed Grimoire priority to reflect the fact it no longer sucks for magic.