Update glacors.txt
Updated AFK guide after Dream sky's feedback, as the method in the guide wasn't fully AFK (it is now 🙏 ), but rather low effort. Also did some minor updates in the rest of the guide and added new glacytes pic. Will update it again later with the grid pic for the cannon placement (it has a palceholder in the mean time).