Update chin.txt
Whoops I forgot to edit the preset.
Removed ornament kits from EoFs, added the specifications of which EoF has which spec, and which arrows were in the quivers to the breakdown itself, although oddly enough it was originally wrong? The note above the preset specified different arrows were in the quivers compared to what the original breakdown said.