Update and rename editing-quick-start.txt to intro-to-editing.txt (#6347)
rewrote this whole channel to be more welcoming. The original file had info on it about roles and points and stuff which I believe belongs in an info-dump channel rather than the channel which should be easy to digest for newcomers. Ideally this channel would get another channel that serves to discuss some more advanced editing stuff like other tools, and that channel would be referenced in the next steps section of this channel.